All programs now share the same `Geometry` and   `display_geometry` implementations Gtk initialization in `dmenu` and `grid` was changed   according to what was done in `bar`
